final ?¡

+++ Kernel/mm/mm.c
@@ -15,7 +15,7 @@ typedef struct MM{
    void * start;
    uint32_t blockQty;
    uint32_t blocksUsed;
-   uint8_t *bitmap;
+   uint32_t *bitmap;
    uint32_t current;
    lastFreed lastFreed;
 } mm;
@@ -50,7 +50,7 @@ int mmInit ( void * baseAddress ,uint64_t memorySize)  {
     return 0;
 }
 static size_t sizeToBlockQty(size_t size){
-    return (size % BLOCK_SIZE)? (size / BLOCK_SIZE) + 1 : size / BLOCK_SIZE;
+    return (size_t) ((size % BLOCK_SIZE)? (size / BLOCK_SIZE) + 1 : size / BLOCK_SIZE);
 }
 
 static void initializeBitmap(){
@@ -58,9 +58,9 @@ static void initializeBitmap(){
         memoryManager.bitmap[i] = FREE;
     }
 }
-static uintptr_t findFreeBlocks( size_t blocksNeeded, int start, int end){
+static uintptr_t findFreeBlocks( size_t blocksNeeded, size_t start, size_t end){
     size_t freeBlocks = 0;
-    for(int i = start; i < end; i++){
+    for(size_t i = start; i < end; i++){
         if(memoryManager.bitmap[i] == FREE){
             freeBlocks++;
             if(freeBlocks == blocksNeeded){
@@ -90,28 +90,29 @@ void * allocMemory(size_t size){
     }
     if(blocksNeeded < memoryManager.lastFreed.blocks){
         return markGroupAsUsed(memoryManager.lastFreed.blocks, memoryManager.lastFreed.index);
+        
     }
     uintptr_t initialBlockAddress = findFreeBlocks( blocksNeeded, memoryManager.current, memoryManager.blockQty);
     
     if(initialBlockAddress == -1){
-     
         initialBlockAddress = findFreeBlocks( blocksNeeded, 0, memoryManager.current + blocksNeeded);
-        
+       
     }
-    if(initialBlockAddress == -1){
+     if(initialBlockAddress == -1){
         
-        return NULL;
-    }
+         return NULL;
+ }
     memoryManager.current = (initialBlockAddress) / BLOCK_SIZE + blocksNeeded;
     if(blocksNeeded == 1){
     
-        memoryManager.blocksUsed++;
+         memoryManager.blocksUsed++;
        
         memoryManager.bitmap[(initialBlockAddress - (uintptr_t) memoryManager.start) / BLOCK_SIZE] = SINGLE_BLOCK;
        
         return (void *) initialBlockAddress;
-    }
-    return markGroupAsUsed(blocksNeeded, (initialBlockAddress - (uintptr_t) memoryManager.start) / BLOCK_SIZE);
+     }
+     return markGroupAsUsed(blocksNeeded, (initialBlockAddress - (uintptr_t) memoryManager.start) / BLOCK_SIZE);
+    
 }
 
 void freeMemory(void * memory){
